+++ configuraciones/paquetes/bajas.js
@@ -1,0 +1,7 @@
+const express = require('express')
+const {body, validationResult} = require('express-validator')
+const router = express.Router()
+const sql = require('../conection')
+
+
+module.exports = router
